+++ pytrobot/__init__.py
@@ -1,9 +1,7 @@
 # pytrobot/__init__.py
 
 # API para o desenvolvedor implementar
-from pytrobot.core import PyTRobot, State, Tool, Transition, First
-from pytrobot.core.states.base_state import BaseState
-from pytrobot.core.tools.base_tool import BaseTool
+from pytrobot.core import PyTRobot, State, Tool, Transition, First, BaseState, BaseTool
 from pytrobot.debug import debug_entrypoint as debug
 from pytrobot.debug import auto_import_classes as auto_import
