+++ pytrobot/core/machine_layer.py
@@ -1,6 +1,4 @@
 # pytrobot/core/machine.py
-from pytrobot.core.states.starter_state import _StarterState
-from pytrobot.core.states.base_state import BaseState
 
 class StateTransitionError(Exception):
     pass
@@ -34,13 +32,16 @@ class TrueTable:
             self._transitions[current_state]['0'] = next_state_on_failure
 
 class StateMachine:
+
     def __init__(self, access_dataset_layer, access_object_layer, access_machine_layer):
+        from pytrobot.core import BaseState
+        from pytrobot.scaffold.src.states.starter_state import _StarterState
         self.access_dataset_layer = access_dataset_layer
         self.access_object_layer = access_object_layer
         self.access_machine_layer = access_machine_layer
-        self.current_state: BaseState = _StarterState()
+        self.current_state : BaseState = _StarterState()
 
-    def get_next_state(self) -> BaseState:
+    def get_next_state(self):
         status = self.current_state._status
         next_state_name = self.access_machine_layer.evaluate_next_state(self.current_state.__class__.__name__, status)
         if not next_state_name:
@@ -76,37 +77,3 @@ class StateMachine:
                     continue
             
             self.current_state = self.access_machine_layer.get_next_state()
-
-# class AccessMachineLayer:
-
-#     def __init__(self, pytrobot_instance):
-#         self.pytrobot_instance = pytrobot_instance
-    
-#     def get_current_state(self):
-#         return self.pytrobot_instance.current_state
-
-#     def add_transition(self, current_state, next_state_on_success, next_state_on_failure):
-#         self.pytrobot_instance.true_table.add_transition(current_state, next_state_on_success, next_state_on_failure)
-
-#     def evaluate_next_state(self, current_state_name, status):
-#         return self.pytrobot_instance.true_table.evaluate_next_state(current_state_name, status)
-    
-#     def get_next_state(self):
-#         return self.pytrobot_instance.state_machine.get_next_state()
-
-#     def reset_current_state(self):
-#         return self.pytrobot_instance.state_machine.reset_current_state()
-
-#     def transition(self, current_state, next_state_on_success=None, next_state_on_failure=None):
-#         """
-#         Update transition in TrueTable.
-        
-#         :param current_state: Current state.
-#         :param next_state_on_success: Next state on success.
-#         :param next_state_on_failure: Next state on failure.
-#         """
-#         self.pytrobot_instance.true_table.update_transition(
-#             current_state, 
-#             next_state_on_success, 
-#             next_state_on_failure
-#         )
