--- pytrobot/core/tools/__init__.py
@@ -1,1 +1,0 @@
-# pytrobot/core/objects_layer/__init__.py
