+++ pytrobot/scaffold/src/__init__.py
@@ -1,4 +1,3 @@
 # DO NOT CHANGE THIS
-import src.actions
 import src.tools
 import src.states
